_blueprint_server/requirements.txt
mcp>=1.0
pydantic>=2.0
pyyaml>=6.0
rich
